auth-service/requirements.txt
fastapi
uvicorn
python-jose[cryptography]
passlib[bcrypt]
psycopg2-binary
python-dotenv
email-validator
